requirements.txt
langgraph==0.4.8
langsmith==0.3.45
python-dotenv==1.1.0
langchain==0.3.25
langchain-openai==0.3.24
openai==1.88.0
langchain_community==0.3.25
django==5.2.3
langgraph-checkpoint-postgres==2.0.21
langchain_tavily==0.2.4